requirements.txt
asgiref==3.8.1
Django==5.0.7
django-guardian==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.8.0
sqlparse==0.5.1
tzdata==2024.1
